+++ src/main/java/ddd/caffeine/ratrip/common/util/YamlPropertySourceFactory.java
@@ -1,7 +1,7 @@
 package ddd.caffeine.ratrip.common.util;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
+import static ddd.caffeine.ratrip.common.exception.ExceptionInformation.*;
+
 import java.util.Properties;
 
 import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
@@ -11,25 +11,26 @@ import org.springframework.core.io.support.EncodedResource;
 import org.springframework.core.io.support.PropertySourceFactory;
 import org.springframework.lang.Nullable;
 
+import ddd.caffeine.ratrip.common.exception.domain.CommonException;
+
 public class YamlPropertySourceFactory implements PropertySourceFactory {
 	@Override
-	public PropertySource<?> createPropertySource(@Nullable String name, EncodedResource resource) throws IOException {
+	public PropertySource<?> createPropertySource(@Nullable String name, EncodedResource resource) {
 		Properties propertiesFromYaml = loadYamlIntoProperties(resource);
-		String sourceName = name != null ? name : resource.getResource().getFilename();
-		return new PropertiesPropertySource(sourceName, propertiesFromYaml);
+		if (name != null) {
+			return new PropertiesPropertySource(name, propertiesFromYaml);
+		}
+		return new PropertiesPropertySource(resource.getResource().getFilename(), propertiesFromYaml);
 	}
 
-	private Properties loadYamlIntoProperties(EncodedResource resource) throws FileNotFoundException {
+	private Properties loadYamlIntoProperties(EncodedResource resource) {
 		try {
 			YamlPropertiesFactoryBean factory = new YamlPropertiesFactoryBean();
 			factory.setResources(resource.getResource());
 			factory.afterPropertiesSet();
 			return factory.getObject();
 		} catch (IllegalStateException e) {
-			Throwable cause = e.getCause();
-			if (cause instanceof FileNotFoundException)
-				throw (FileNotFoundException)e.getCause();
-			throw e;
+			throw new CommonException(ILLEGAL_YML_PROPERTIES_EXCEPTION);
 		}
 	}
 }
